--- TercerMiniProyecto/src/App.java
@@ -1,12 +1,0 @@
-import modelo.Juego;
-import vista.VistaJuegoConsola;
-import vista.VistaJuegoGUI;
-
-public class App {
-    public static void main(String[] args) throws Exception {
-        Juego modelo = new Juego();
-        VistaJuegoConsola vista = new VistaJuegoConsola();
-        //VistaJuegoGUI vista = new VistaJuegoGUI();
-
-    }
-}
